requirements.txt
pillow
argparse
SimpleITK
numpy
psutil
